Self-trigger: offset decimali, e niente segfault sui valori non interi
La soglia resta intera perche' il registro 0x1n80 e' a 12 bit, ma
l'offset ora e' un numero reale: 'soglia = baseline - offset' con la
baseline gia' frazionaria, quindi un offset decimale sceglie su quale
intero si arrotonda.

Soprattutto toglie due trappole. Nel TOML una lista con un valore
decimale finiva in Config::GetEntryList, che chiama get_as<int64_t>() su
un nodo float, ottiene nullptr e lo dereferenzia: segfault. Ora l'array
viene letto con value<double>(), che converte anche gli interi, e un
elemento non numerico da' un errore esplicito. Nel file di comando a
runtime '8 15.5' veniva troncato a 15 in silenzio; ora i decimali si
leggono e una riga malformata viene segnalata invece che interpretata a
meta'.

+++ src/Digitizer.h
@@ -135,8 +135,8 @@ private:
     bool fSelfTriggerRelative;
     std::vector<uint32_t> fSelfTriggerChannels;
     std::map<uint32_t,uint32_t> fSelfTriggerThreshold;   // per canale, 12 bit (0..4095)
-    uint32_t fSelfTriggerThresholdOffset;                // conteggi ADC, modo "relative"
-    std::map<uint32_t,uint32_t> fSelfTriggerOffset;      // stesso, ma per canale
+    double fSelfTriggerThresholdOffset;                  // conteggi ADC, modo "relative"
+    std::map<uint32_t,double> fSelfTriggerOffset;        // stesso, ma per canale
     std::map<uint32_t,double> fTransparentBaseline;      // baseline in Transparent Mode
     std::map<uint32_t,double> fTransparentRMS;
